+++ bigbluebutton-html5/imports/api/audio/client/bridge/base.js
@@ -1,3 +1,14 @@
+import { Tracker } from 'meteor/tracker';
+import VoiceCallStates from '/imports/api/voice-call-states';
+import CallStateOptions from '/imports/api/voice-call-states/utils/callStates';
+import logger from '/imports/startup/client/logger';
+import Auth from '/imports/ui/services/auth';
+
+const MEDIA = Meteor.settings.public.media;
+const BASE_BRIDGE_NAME = 'base';
+const CALL_TRANSFER_TIMEOUT = MEDIA.callTransferTimeout;
+const TRANSFER_TONE = '1';
+
 export default class BaseAudioBridge {
   constructor(userData) {
     this.userData = userData;
@@ -19,6 +30,8 @@ export default class BaseAudioBridge {
       reconnecting: 'reconnecting',
       autoplayBlocked: 'autoplayBlocked',
     };
+
+    this.bridgeName = BASE_BRIDGE_NAME;
   }
 
   getPeerConnection() {
@@ -39,5 +52,51 @@ export default class BaseAudioBridge {
 
   changeOutputDevice() {
     console.error('The Bridge must implement changeOutputDevice');
+  }
+
+  sendDtmf() {
+    console.error('The Bridge must implement sendDtmf');
+  }
+
+  trackTransferState(transferCallback) {
+    return new Promise((resolve, reject) => {
+      let trackerControl = null;
+
+      const timeout = setTimeout(() => {
+        trackerControl.stop();
+        logger.warn({ logCode: 'audio_transfer_timed_out' },
+          'Timeout on transferring from echo test to conference');
+        this.callback({
+          status: this.baseCallStates.failed,
+          error: 1008,
+          bridgeError: 'Timeout on call transfer',
+          bridge: this.bridgeName,
+        });
+
+        this.exitAudio();
+
+        reject(this.baseErrorCodes.REQUEST_TIMEOUT);
+      }, CALL_TRANSFER_TIMEOUT);
+
+      this.sendDtmf(TRANSFER_TONE);
+
+      Tracker.autorun((c) => {
+        trackerControl = c;
+        const selector = { meetingId: Auth.meetingID, userId: Auth.userID };
+        const query = VoiceCallStates.find(selector);
+
+        query.observeChanges({
+          changed: (id, fields) => {
+            if (fields.callState === CallStateOptions.IN_CONFERENCE) {
+              clearTimeout(timeout);
+              transferCallback();
+
+              c.stop();
+              resolve();
+            }
+          },
+        });
+      });
+    });
   }
 }
